streamlined content deletion when all files are unwanted

+++ code/Domain/Models/Arr/Queue/QueueRecord.cs
@@ -3,17 +3,17 @@
 public record QueueRecord
 {
     // Sonarr
-    public int SeriesId { get; init; }
-    public int EpisodeId { get; init; }
-    public int SeasonNumber { get; init; }
+    public long SeriesId { get; init; }
+    public long EpisodeId { get; init; }
+    public long SeasonNumber { get; init; }
     
     // Radarr
-    public int MovieId { get; init; }
+    public long MovieId { get; init; }
     
     // Lidarr
-    public int ArtistId { get; init; }
+    public long ArtistId { get; init; }
     
-    public int AlbumId { get; init; }
+    public long AlbumId { get; init; }
     
     // common
     public required string Title { get; init; }
@@ -22,5 +22,5 @@ public record QueueRecord
     public string TrackedDownloadState { get; init; }
     public required string DownloadId { get; init; }
     public required string Protocol { get; init; }
-    public required int Id { get; init; }
+    public required long Id { get; init; }
 }
